Move dash models to a separate package
